requirements.txt
# Core Python Dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
fastapi-cors==0.0.6
fastapi-swagger-ui==0.1.0

# HTTP Client (Pinned for FastAPI/Starlette compatibility)
httpx==0.22.0

# Database
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6

# File Parsing
PyPDF2==3.0.1
python-docx==0.8.11

# OCR Support
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3

# Natural Language Processing
spacy==3.5.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
nltk==3.8.1
transformers==4.30.2
torch==2.0.1

# PDF Report Generation
reportlab==3.6.12

# Multipart Form Data Support
python-multipart==0.0.6

# Development and Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Machine Learning and Data Processing
scikit-learn==1.2.2
numpy==1.26.0
pandas==2.0.2

# Logging and Monitoring
loguru==0.7.0

# Type Hints
typing==3.7.4.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Add requests for Hugging Face API calls
requests

# Added from the code block
matplotlib==3.8.0
plotly==6.0.1